refactor(web): split api/manifest.ts and mcp/tools.ts into directories
Both files crossed the 350-LOC project rule after the policy-moderator
slices added a manifest entry, an MCP tool, and routing branches:

  api/manifest.ts: 452 → 464 (my +12) → split into
    types.ts        — EndpointId / McpToolName unions + spec shapes
    endpoints.ts    — ENDPOINTS array + endpointSpec helper
    mcp-tools.ts    — MCP_TOOLS array + helpers + boot-time invariants
    index.ts        — barrel

  mcp/tools.ts: 620 → 658 (my +38) → split into per-domain
    tools/helpers.ts          — textResult, formatZodIssues
    tools/submission-write.ts — submit_link, update_submission, delete_submission
    tools/comment-write.ts    — post_comment, update_comment, delete_comment
    tools/vote-save.ts        — vote, save
    tools/notification.ts     — list_notifications, mark_notifications_read
    tools/identity.ts         — get_constitution, get_quota, list_my_decisions, me
    tools/index.ts            — composes all of the above + registerReadTools

read-tools.ts (the 12 read tools) stays at lib/mcp/read-tools.ts;
the new tools/index.ts pulls it in via the existing relative import.

The api-manifest drift test (tests/api-manifest.test.ts) walks the
new tools/ directory at discovery time so additions land under the
drift check automatically. read-tools.ts still walked explicitly.

Two new MCP tool branches surface from this work: submit_link now
maps the 'rate' SubmitResult reason (Slice 3 ladder), and
post_comment maps both 'illegal' and 'rate' CommentResult reasons
(Slices 2 + 3). Both produce LLM-readable error text for clients
hitting the policy moderator.

41 importers of @/db/schema, 41 of @/lib/api/manifest, and 4 of
@/lib/comments/@/lib/submissions resolve through their barrels
unchanged. No behavioral change.
