+++ web/src/lib/mcp/tools/helpers.ts
@@ -1,0 +1,17 @@
+/**
+ * Tiny helpers shared across the per-domain MCP tool modules.
+ */
+
+import type { z } from "zod";
+
+export function textResult(text: string, isError = false) {
+  return { isError, content: [{ type: "text" as const, text }] };
+}
+
+// Format a Zod safeParse error into the inline text the tool result
+// surfaces back to the LLM client.
+export function formatZodIssues(error: z.ZodError): string {
+  return error.issues
+    .map((i) => `${i.path.join(".") || "<root>"}: ${i.message}`)
+    .join("; ");
+}
